pyproject.toml
[tool.poetry]
name = "strava-peaks"
version = "0.1.0"
description = ""
authors = ["Adam Saegebarth <adam.saegebarth@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
pyyaml = "^5.3.1"
boto3 = "^1.14.0"
stravalib = "^0.10.2"
python-jose = "^3.1.0"
jinja2 = "^2.11.2"
ratelimiter = "^1.2.0"
aws-xray-sdk = "^2.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.5.3"
black = {version = "^19.10b0", allow-prereleases = true}
cfn-lint = "^0.33.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
